ref: split tests, setup and ChainSelectors
