+++ contracts/tests/utils/Selectors.ts
@@ -1,0 +1,9 @@
+export const ChainSelectors = {
+  testnet: {
+    ton: 13879075125137744094n, // ton-localnet
+  },
+  testselectors: {
+    CHAINSEL_EVM_TEST_90000001: 909606746561742123n,
+    CHAINSEL_EVM_TEST_90000002: 5548718428018410741n,
+  },
+}
